requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
DateTime==5.2
Flask==3.0.0
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
mysqlx==2.1.0
packaging==23.2
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.2
sqlx-batis==2.1.0
sqlx-exec==2.2.2
Werkzeug==3.0.1
zope.interface==6.1
